requirements.txt
# Core LLM and API clients
google-generativeai>=0.8.0
openai>=1.0.0
google-api-python-client>=2.100.0
google-cloud-vision>=3.10.0
google-auth>=2.20.0
google-auth-httplib2>=0.2.0
google-api-core>=2.15.0
google-cloud-aiplatform>=1.50.0 # For Vertex AI
vertexai>=1.50.0 # Explicitly add vertexai

# Image processing and data handling
Pillow>=10.4.0
numpy>=1.26.0
PyYAML>=6.0.1
requests>=2.20.0 # For Stability AI and other HTTP requests
opencv-python>=4.8.0 # Optional, if needed beyond Pillow
ipython>=7.0.0 # For IPython.display and other utilities

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging
python-json-logger>=2.0.7

# Font support
fonttools>=4.50.0

# Other
python-dotenv